Update PatchMW2T6Clips.cpp
Removed NTA's CoD:OL weapon testing code, as well as updated the attachment code to support any peacekeeper variant. (Seeing as the .FF for it [weap_trey] is not included, neither of these changes should really make a difference. But, if they are, I might take a look into this code again ;)